+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## next
 
+- Fixed `stringifyStream()` and `stringifyInfo()` to work properly when replacer is an allowlist
 - `parseChunked()`
     - Fixed wrong parse error when chunks are splitted on a whitespace inside an object or array (#6, @alexei-vedder)
     - Fixed corner cases when wrong placed or missed comma doesn't cause to parsing failure